+++ tests/testthat/test_params.R
@@ -32,3 +32,11 @@ test_that("cardiac_output works", {
         tolerance = 5e-2
     )
 })
+
+test_that("partial_pressures works", {
+    expect_error(partial_pressures(), "pinsp")
+    expect_identical(
+        partial_pressures(pinsp = 12),
+        c(pinsp = 12, lung = 0, vrg = 0, mus = 0, fat = 0, cv = 0)
+    )
+})