+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Changes in development version
 
+- rename tissue_coefficients to partition_coefficients
+
 ## Changes in 0.0.2
 
 - Add humidification.