+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes in development version
 
-- rename tissue_coefficients to partition_coefficients
+- Add `partial_pressures` helper function.
+- Rename `tissue_coefficients` to `partition_coefficients`.
 
 ## Changes in 0.0.2