+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## Changes in development version
 
+## Changes in 0.0.3
+
+- Add metabolism support. Second decimal place doesn't completely match the data
+  in Cowles 1973.
 - Add pulmonary shunt support.
 - Add `partial_pressures` helper function.
 - Rename `tissue_coefficients` to `partition_coefficients`.